requirements.txt
requests
pyyaml
fdroidserver
gitpython
qrcode
yamllint
Pillow
androguard==3.4.0